Create daffodil-runtime1-layers module and dynamic SPI loader of layer defs.
daffodil-runtime1-layers holds all the built-in layer implementations
that Daffodil provides.

Note that the AISPayloadArmoring, and some other layering
implementations (IPv4, CheckDigits) are partial and are just in
daffodil-test, since they are just to provide test coverage. They're
incomplete, and not supported as usable layer transforms, just tests/examples.
They do illustrate how to do checksums/check-digits, and can be
generalized to create a real IP checksums layered schema.

Layers now are defined with a LayerCompiler that does schema-compile-time
checking and returns the LayerTransformerFactory object (which is
the serialized object).

Unsolved mystery: I don't understand why changes were required to two tests:

testParseLineFoldIMF2
testUnparseLineFoldIMF2

The changes I have made should not have made any functional difference.
However, these two tests were incorrect. IMF line folding should
not remove embedded \r\n unless it is followed by a space. Both these
tests were depending on that behavior.

DAFFODIL-1927

+++ daffodil-io/src/main/scala/org/apache/daffodil/io/LimitingJavaIinputStreams.scala
@@ -36,7 +36,7 @@ import java.nio.charset.StandardCharsets
  * Thread safety: This is inherently stateful - so not thread safe to use
  * this object from more than one thread.
  */
-class ExplicitLengthLimitingStream(in: InputStream, limit: Int)
+class ExplicitLengthLimitingStream(in: InputStream, limit: Long)
   extends FilterInputStream(in) {
 
   private var numRemaining = limit
@@ -46,7 +46,7 @@ class ExplicitLengthLimitingStream(in: InputStream, limit: Int)
     if (numRemaining == 0) -1
     else if (len == 0) 0
     else {
-      val requestSize = math.min(numRemaining, len)
+      val requestSize = math.min(numRemaining, len).toInt
       val actualSize = in.read(buf, off, requestSize)
       if (actualSize == -1)
         numRemaining = 0
